requirements.txt
Django==4.1.2
pillow
django-rest-framework